+++ test/e2e/specs/field-render-tests.js
@@ -96,6 +96,7 @@ module.exports = {
     browser.expect.element('#date').to.be.visible
 
     browser.click('#date')
+    browser.pause(500)
     browser.expect.element('.flatpickr-calendar').to.be.visible
 
     browser.click('.today')
@@ -123,9 +124,9 @@ module.exports = {
   'Fill out date time field using picker': function (browser) {
     browser.options.desiredCapabilities.name = 'Fill out date time field using picker'
 
-    browser.expect.element('#date_time').to.be.visible
+    browser.expect.element('fieldset#date_time-fs').to.be.visible
 
-    browser.click('#date_time')
+    browser.click('fieldset#date_time-fs')
     browser.expect.element('body > div.flatpickr-calendar.hasTime.animate.open').to.be.visible
 
     browser.click('body > div.flatpickr-calendar.hasTime.animate.open > div.flatpickr-innerContainer > div > div.flatpickr-days > div > span.flatpickr-day.today')